Consolidate frontend config into a single file
Move GOOGLE_CLIENT_ID from src/frontend/auth/config.ts into
src/frontend/config.ts and delete the auth-local file. Splitting a
single constant into its own module by responsibility was overkill.
