+++ src/frontend/app.tsx
@@ -1,6 +1,6 @@
 import { GoogleOAuthProvider } from '@react-oauth/google';
 import { createRoot } from 'react-dom/client';
-import { GOOGLE_CLIENT_ID } from './auth/config';
+import { GOOGLE_CLIENT_ID } from './config';
 import { RoadStationMap } from './components/RoadStationMap';
 
 const container = document.getElementById('map-canvas');
