--- src/frontend/auth/config.ts
@@ -1,4 +1,0 @@
-// Public OAuth client ID for Google Identity Services. The security gate is the
-// Authorized JavaScript origin configured in the Google Cloud Console, not the
-// secrecy of this value.
-export const GOOGLE_CLIENT_ID = '740963820239-n5t81ueg3b27mtcs229fk62na8bfjvva.apps.googleusercontent.com';
